requirements.txt
Django==3.2.17
psycopg2-binary==2.9.5  
Pillow==8.4.0
debugpy==1.5.1
redis==4.3.0
django-redis==5.2.0
